# docs/api_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Equity Lab",
    "description": "Institutional Indian Equity Research, Options Arbitrage, and Return Probability Engine",
    "version": "0.0.9"
  },
  "paths": {
...
    "/api/v1/strategies": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Fetch All Strategies",
        "description": "Returns list of all strategy modules and production status.",
        "operationId": "fetch_all_strategies_api_v1_strategies_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyModule"
                  },
                  "title": "Response Fetch All Strategies Api V1 Strategies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/strategies/swing-alerts": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Fetch Swing Trade Alerts",
        "description": "Returns active short-to-medium term high-probability swing trade alerts.",
        "operationId": "fetch_swing_trade_alerts_api_v1_strategies_swing_alerts_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional symbol to filter",
              "title": "Symbol"
            },
            "description": "Optional symbol to filter"
          },
          {
            "name": "universe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional universe descriptor",
              "title": "Universe"
            },
            "description": "Optional universe descriptor"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/strategies/{strategy_id}": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Fetch Strategy Detail",
        "description": "Returns details for a specific strategy module.",
        "operationId": "fetch_strategy_detail_api_v1_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyModule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/strategies/{strategy_id}/run": {
      "post": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Strategy",
        "description": "Runs screening or diagnostic analysis for a specific strategy module.",
        "operationId": "run_strategy_api_v1_strategies__strategy_id__run_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/growth-inflection": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Growth Inflection",
        "description": "Executes Strategy E1: Growth Inflection Engine (Revenue, Profit, EPS, Margin, ROCE, FCF acceleration).",
        "operationId": "run_growth_inflection_api_v1_research_growth_inflection_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock symbol (e.g. RELIANCE)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthInflectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/turnaround-stage": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Turnaround Stage",
        "description": "Executes Strategy E2: Turnaround Stage Engine (Distress to Recovery lifecycle & False Turnaround detection).",
        "operationId": "run_turnaround_stage_api_v1_research_turnaround_stage_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock symbol (e.g. RELIANCE)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnaroundStageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/growth-market-gap": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Growth Market Gap",
        "description": "Executes Strategy E3: Growth vs Market Recognition Gap Engine (Business growth CAGR vs Stock price CAGR).",
        "operationId": "run_growth_market_gap_api_v1_research_growth_market_gap_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock symbol (e.g. RELIANCE)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthMarketGapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/governance-quality": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Governance Quality",
        "description": "Evaluates Governance & Management Quality (Promoter pledge, CFO/PAT hygiene, governance events).",
        "operationId": "run_governance_quality_api_v1_research_governance_quality_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock symbol (e.g. RELIANCE)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernanceQualityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/research/growth-arbitrage": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Run Growth Arbitrage",
        "description": "Executes Strategy E5: AI Growth Arbitrage & DCF Valuation Engine (Institutional Grade).",
        "operationId": "run_growth_arbitrage_api_v1_research_growth_arbitrage_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock symbol (e.g. RELIANCE)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthArbitrageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/scorecard": {
      "get": {
        "tags": [
          "Expert Strategy & Research Engines"
        ],
        "summary": "Fetch Symbol Scorecard",
        "description": "Returns consolidated Scorecard item combining Arbiter, Multibagger, and Return Probabilities.",
        "operationId": "fetch_symbol_scorecard_api_v1_research_scorecard_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stock ticker symbol (e.g. RELIANCE)",
              "title": "Symbol"
            },
            "description": "Stock ticker symbol (e.g. RELIANCE)"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyModule": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "required_inputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Inputs"
          },
          "universe": {
            "type": "string",
            "title": "Universe"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metrics"
          },
          "risk_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Warnings"
          },
          "methodology": {
            "type": "string",
            "title": "Methodology"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "description",
          "status",
          "required_inputs",
          "universe",
          "metrics",
          "risk_warnings",
          "methodology"
        ],
        "title": "StrategyModule"
      },
...
    }
  }
}